refactor(core): move secrets + frames test modules into sibling files (<1000)
secrets/mod.rs (1067) and gateway/protocol/frames.rs (1298) were dominated by
large #[cfg(test)] modules. Move them to sibling test files:

- secrets/tests.rs
- gateway/protocol/frames/{tests.rs, frame_size_tests.rs}

secrets/mod.rs → 144 lines; frames.rs → 637 lines. No behavior change.
